+++ src/lib/perfect-scrollbar.classes.ts
@@ -1,0 +1,7 @@
+export class Geometry {
+  x: number;
+  y: number;
+
+  w: number;
+  h: number;
+}